+++ packages/viewer/web/src/RichContentViewer.jsx
@@ -8,6 +8,7 @@ import {
   getLangDir,
   SPOILER_TYPE,
   GlobalContext,
+  Version,
 } from 'wix-rich-content-common';
 import 'wix-rich-content-common/dist/statics/styles/draftDefault.rtlignore.scss';
 import { convertToReact } from './utils/convertContentState';
@@ -15,6 +16,7 @@ import viewerStyles from '../statics/rich-content-viewer.scss';
 import viewerAlignmentStyles from '../statics/rich-content-viewer-alignment.rtlignore.scss';
 import rtlStyle from '../statics/rich-content-viewer-rtl.rtlignore.scss';
 import { deprecateHelpers } from 'wix-rich-content-common/dist/lib/deprecateHelpers.cjs.js';
+import { combineMappers } from './utils/combineMappers';
 
 class RichContentViewer extends Component {
   constructor(props) {
@@ -24,6 +26,7 @@ class RichContentViewer extends Component {
     this.state = {
       raw: {},
     };
+    this.typeMappers = combineMappers(props.typeMappers);
   }
 
   static getInitialState = props => {
@@ -84,6 +87,20 @@ class RichContentViewer extends Component {
     return { error };
   }
 
+  componentDidMount() {
+    import(
+      /* webpackChunkName: debugging-info */ 'wix-rich-content-common/lib/debugging-info'
+    ).then(({ reportDebuggingInfo }) => {
+      reportDebuggingInfo({
+        version: Version.currentVersion,
+        reporter: 'Rich Content Viewer',
+        plugins: Object.keys(this.typeMappers),
+        getContent: () => this.props.initialState,
+        getConfig: () => this.props.config,
+      });
+    });
+  }
+
   render() {
     const { onError, config = {} } = this.props;
     try {
@@ -94,7 +111,6 @@ class RichContentViewer extends Component {
       const { styles } = this;
       const {
         textDirection,
-        typeMappers,
         decorators,
         inlineStyleMappers,
         locale,
@@ -114,7 +130,7 @@ class RichContentViewer extends Component {
       const output = convertToReact(
         styles,
         textDirection,
-        typeMappers,
+        this.typeMappers,
         contextualData,
         decorators,
         inlineStyleMappers,